+++ playwright.config.js
@@ -28,6 +28,10 @@ module.exports = defineConfig({
     {
       name: 'mobile-chrome',
       use: {...devices['Pixel 7']}
+    },
+    {
+      name: 'mobile-safari',
+      use: {...devices['iPhone 13'], browserName: 'webkit'}
     }
   ]
 });